+++ test/middleware/init.test.js
@@ -10,12 +10,14 @@ describe('middleware/init', function() {
     chai.express.use(function(req, res, next) {
       expect(res.status).to.be.a('function');
       expect(res.send).to.be.a('function');
+      expect(res.json).to.be.a('function');
       expect(res.location).to.be.a('function');
       res.end();
     })
     .request(function(req, res) {
       expect(res.status).to.be.undefined;
       expect(res.send).to.be.undefined;
+      expect(res.json).to.be.undefined;
       expect(res.location).to.be.undefined;
     })
     .finish(done)