+++ CHANGELOG.md
@@ -39,8 +39,9 @@ rather than just having a `statusCode` property.
 `http.ServerResponse`.
 - Changed `Response#redirect` signature from `function(url, status)` to
 `function(status, url)`, matching Express 4.x API.
-- `Response#status`, `Response.send`, and `Response#location` added to `res` in
-`init()` middleware rather than being part of base `Response` prototype.
+- `Response#status`, `Response#send`, `Response#json` and `Response#location`
+added to `res` in `init()` middleware rather than being part of base `Response`
+prototype.
 
 ### Removed